LPS-52122 Tests SF
