+++ portal-test/src/com/liferay/portlet/dynamicdatalists/util/test/DDLRecordTestUtil.java
@@ -14,117 +14,15 @@
 
 package com.liferay.portlet.dynamicdatalists.util.test;
 
-import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.util.test.TestPropsValues;
-import com.liferay.portlet.dynamicdatamapping.model.DDMForm;
-import com.liferay.portlet.dynamicdatamapping.model.DDMFormField;
-import com.liferay.portlet.dynamicdatamapping.model.LocalizedValue;
-import com.liferay.portlet.dynamicdatamapping.model.UnlocalizedValue;
-import com.liferay.portlet.dynamicdatamapping.model.Value;
-import com.liferay.portlet.dynamicdatamapping.storage.DDMFormFieldValue;
-import com.liferay.portlet.dynamicdatamapping.storage.DDMFormValues;
-
-import java.util.LinkedHashSet;
-import java.util.Locale;
-import java.util.Set;
 
 /**
  * @author Marcellus Tavares
  * @author André de Oliveira
  */
 public class DDLRecordTestUtil {
-
-	public static Set<Locale> createAvailableLocales(Locale... locales) {
-		Set<Locale> availableLocales = new LinkedHashSet<>();
-
-		for (Locale locale : locales) {
-			availableLocales.add(locale);
-		}
-
-		return availableLocales;
-	}
-
-	public static DDMForm createDDMForm(
-		Set<Locale> availableLocales, Locale defaultLocale) {
-
-		DDMForm ddmForm = new DDMForm();
-
-		ddmForm.setAvailableLocales(availableLocales);
-		ddmForm.setDefaultLocale(defaultLocale);
-
-		return ddmForm;
-	}
-
-	public static DDMFormField createDDMFormField(
-		String name, String label, String type, String dataType,
-		boolean localizable, boolean repeatable, boolean required) {
-
-		DDMFormField ddmFormField = new DDMFormField(name, type);
-
-		ddmFormField.setDataType(dataType);
-		ddmFormField.setLocalizable(localizable);
-		ddmFormField.setRepeatable(repeatable);
-		ddmFormField.setRequired(required);
-
-		LocalizedValue localizedValue = ddmFormField.getLabel();
-
-		localizedValue.addString(LocaleUtil.US, label);
-
-		return ddmFormField;
-	}
-
-	public static DDMFormFieldValue createDDMFormFieldValue(
-		String name, Value value) {
-
-		DDMFormFieldValue ddmFormFieldValue = new DDMFormFieldValue();
-
-		ddmFormFieldValue.setInstanceId(StringUtil.randomString());
-		ddmFormFieldValue.setName(name);
-		ddmFormFieldValue.setValue(value);
-
-		return ddmFormFieldValue;
-	}
-
-	public static DDMFormValues createDDMFormValues(
-		DDMForm ddmForm, Set<Locale> availableLocales, Locale defaultLocale) {
-
-		DDMFormValues ddmFormValues = new DDMFormValues(ddmForm);
-
-		ddmFormValues.setAvailableLocales(availableLocales);
-		ddmFormValues.setDefaultLocale(defaultLocale);
-
-		return ddmFormValues;
-	}
-
-	public static DDMFormFieldValue createLocalizedTextDDMFormFieldValue(
-		String name, String enValue) {
-
-		Value localizedValue = new LocalizedValue(LocaleUtil.US);
-
-		localizedValue.addString(LocaleUtil.US, enValue);
-
-		return createTextDDMFormFieldValue(name, localizedValue);
-	}
-
-	public static DDMFormFieldValue createTextDDMFormFieldValue(
-		String name, Value value) {
-
-		DDMFormFieldValue ddmFormFieldValue = new DDMFormFieldValue();
-
-		ddmFormFieldValue.setInstanceId(StringUtil.randomString());
-		ddmFormFieldValue.setName(name);
-		ddmFormFieldValue.setValue(value);
-
-		return ddmFormFieldValue;
-	}
-
-	public static DDMFormFieldValue createUnlocalizedTextDDMFormFieldValue(
-		String name, String value) {
-
-		return createTextDDMFormFieldValue(name, new UnlocalizedValue(value));
-	}
 
 	public static String getBasePath() {
 		return "com/liferay/portlet/dynamicdatalists/dependencies/";
