+++ portal-test/src/com/liferay/portlet/dynamicdatamapping/util/test/DDMStructureTestHelper.java
@@ -15,7 +15,6 @@
 package com.liferay.portlet.dynamicdatamapping.util.test;
 
 import com.liferay.portal.kernel.locale.test.LocaleTestUtil;
-import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.util.PortalUtil;
@@ -25,20 +24,13 @@ import com.liferay.portlet.dynamicdatalists.model.DDLRecordSet;
 import com.liferay.portlet.dynamicdatalists.util.test.DDLRecordTestUtil;
 import com.liferay.portlet.dynamicdatamapping.io.DDMFormXSDDeserializerUtil;
 import com.liferay.portlet.dynamicdatamapping.model.DDMForm;
-import com.liferay.portlet.dynamicdatamapping.model.DDMFormField;
 import com.liferay.portlet.dynamicdatamapping.model.DDMFormLayout;
 import com.liferay.portlet.dynamicdatamapping.model.DDMStructure;
 import com.liferay.portlet.dynamicdatamapping.model.DDMStructureConstants;
-import com.liferay.portlet.dynamicdatamapping.model.LocalizedValue;
 import com.liferay.portlet.dynamicdatamapping.service.DDMStructureLocalServiceUtil;
 import com.liferay.portlet.dynamicdatamapping.storage.StorageType;
 import com.liferay.portlet.dynamicdatamapping.util.DDMUtil;
 import com.liferay.portlet.dynamicdatamapping.util.DDMXMLUtil;
-
-import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Locale;
-import java.util.Set;
 
 /**
  * @author Eduardo Garcia
@@ -105,66 +97,6 @@ public class DDMStructureTestHelper {
 			testClass, "test-structure.xsd");
 
 		return addStructure(definition, StorageType.JSON.toString());
-	}
-
-	public void addTextDDMFormFields(DDMForm ddmForm, String... fieldNames) {
-		List<DDMFormField> ddmFormFields = ddmForm.getDDMFormFields();
-
-		for (String fieldName : fieldNames) {
-			ddmFormFields.add(createTextDDMFormField(fieldName));
-		}
-	}
-
-	public Set<Locale> createAvailableLocales(Locale... locales) {
-		Set<Locale> availableLocales = new LinkedHashSet<>();
-
-		for (Locale locale : locales) {
-			availableLocales.add(locale);
-		}
-
-		return availableLocales;
-	}
-
-	public DDMForm createDDMForm(
-		Set<Locale> availableLocales, Locale defaultLocale) {
-
-		DDMForm ddmForm = new DDMForm();
-
-		ddmForm.setAvailableLocales(availableLocales);
-		ddmForm.setDefaultLocale(defaultLocale);
-
-		return ddmForm;
-	}
-
-	public DDMForm createDDMForm(String... fieldNames) {
-		DDMForm ddmForm = createDDMForm(
-			createAvailableLocales(LocaleUtil.US), LocaleUtil.US);
-
-		addTextDDMFormFields(ddmForm, fieldNames);
-
-		return ddmForm;
-	}
-
-	public DDMFormField createTextDDMFormField(String name) {
-		return createTextDDMFormField(name, true, false, false);
-	}
-
-	public DDMFormField createTextDDMFormField(
-		String name, boolean localizable, boolean repeatable,
-		boolean required) {
-
-		DDMFormField ddmFormField = new DDMFormField(name, "text");
-
-		ddmFormField.setDataType("string");
-		ddmFormField.setLocalizable(localizable);
-		ddmFormField.setRepeatable(repeatable);
-		ddmFormField.setRequired(required);
-
-		LocalizedValue localizedValue = ddmFormField.getLabel();
-
-		localizedValue.addString(LocaleUtil.US, name);
-
-		return ddmFormField;
 	}
 
 	public DDMForm toDDMForm(String definition) throws Exception {
